#80: Use size_t for Driver::Options channel and buffer fields
- Driver::Options: change numInputs, numOutputs, bufferSize from
  optional<int> to optional<size_t> — consistent with the size_t types
  used throughout the codebase for counts and sizes
- Methcla_AudioDriverOptions (internal C++ struct, opaque from C API):
  replace int fields with -1 sentinel with optional<size_t>; eliminates
  the need for sentinel values and avoids narrowing conversions
- RtAudioDriver: update channel/buffer arithmetic to unsigned int /
  size_t with safe guarded casts (std::max(0u, ...) before converting
  from int inputChannels)

+++ src/Methcla/Audio/IO/Driver.hpp
@@ -32,9 +32,9 @@ namespace Methcla { namespace Audio { namespace IO {
         struct Options
         {
             std::optional<double> sampleRate;
-            std::optional<int>    numInputs;
-            std::optional<int>    numOutputs;
-            std::optional<int>    bufferSize;
+            std::optional<size_t> numInputs;
+            std::optional<size_t> numOutputs;
+            std::optional<size_t> bufferSize;
         };
 
         typedef void (*ProcessCallback)(void* data, Methcla_Time currentTime,
